+++ framework/hyper-express/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @forklaunch/hyper-express
 
+## 1.2.32
+
+### Patch Changes
+
+- update dependency versions
+- Updated dependencies
+  - @forklaunch/validator@1.2.18
+  - @forklaunch/common@1.2.18
+  - @forklaunch/core@1.5.2
+  - @forklaunch/ws@1.2.32
+
 ## 1.2.31
 
 ### Patch Changes